+++ src/__tests__/Seq.spec.ts
@@ -460,7 +460,7 @@ describe("Seq", () => {
 
   describe("chain", () => {
     test("should map the current seq", () => {
-      const result = fromArray([1, 2, 3, 4]).chain(seq => seq.toArray());
+      const result = fromArray([1, 2, 3, 4]).pipe(seq => seq.toArray());
 
       expect(result).toEqual([1, 2, 3, 4]);
     });